Fix rsrc with bad RVA.
In some cases the resource data has an RVA which does not map to a file offset.
Prior to this change we would stop parsing that resource when this is found,
which means we are losing useful information like the length and type/name
strings (if they exist).

I checked how tools like pe-studio and CFF explorer handle this case and they
just display the RVA with an unknown offset.

I think it is useful to not stop parsing if the RVA does not map to a file
offset. This commit makes it so "invalid" RVAs are set as YR_UNDEFINED and we
continue parsing. I'm also adding support for setting the RVA, because it is
useful to know in some cases.

I was careful to verify that in the case of parsing version information
resources that function does check that the RVA maps to a legitimate file offset
before parsing, so this is safe to do.