Fix regex parsing on aarch64.
Turns out that on aarch64 (and other similar architectures) tests are failing
(see the backtrace below) when trying to parse an invalid escape sequence. To
fix this I've changed yr_parse_re_string() to take a length field and use
yy_scan_bytes() instead of yy_scan_string(). This is because when
read_escaped_char() is eventually called it assumes there will be 3 bytes to
read (xAA, for example). The test case is trying to get it to parse "x" with no
hex digits, so it ends up reading off the end of the string causing a segfault
on aarch64. This doesn't trigger on other x86/x64 architectures probably due to
some weird MMU thing.

Program received signal SIGSEGV, Segmentation fault.
0xb6bebd90 in __GI__IO_fread (buf=0xc4410, size=1, count=1, fp=0x0) at iofread.c:37
37      iofread.c: No such file or directory.
(gdb) bt
    str=0xca4e0, string_ref=0xbeffd6e0) at parser.c:714
    compiler=0xc5ac8) at lexer.l:901
    rules_string=0x8d720 "rule test { strings: $a = /\\x/ condition: $a }", namespace_=0x0) at compiler.c:696
    at tests/util.c:140
(gdb)

Fixes #1344.