Do not directly include stdbool.h (#1377)
There's a special handling for bool type in utils.h which checks that
stdbool.h is really in the system and uses it, or it defines its own
bool type. All uses of bool type should go through there instead of
direct include of stdbool.h as it can cause issues with compilation
units having different sizes of bool type if they are mixed together.