Make yr_scanner_scan_mem_blocks easier to follow.
The main change is that *every* call to the iterator's "first" and "next" functions must set the "last_error" field, to either ERROR_SUCCESS or some other error (possibly ERROR_BLOCK_NOT_READY). Also "last_error == ERROR_BLOCK_NOT_READY" is used for detecting whether or not yr_scanner_scan_mem_blocks is being called for the first time or it's a retry, instead of using "matches_notebook == NULL" which is more confusing.