Refactor the way in which memory block iterators set the file size.
Now memory block iterator must implement a "file_size" function (in addition to "first" and "next"). This function is always invoked when all blocks has been iterated, and the iterator implementation is responsible for returning the correct value for the file size. This way the only thing you need to do for implementing a memory block iterator is providing these three functions. You don't need to rely on any fields of YR_MEMORY_BLOCK_ITERATOR, which except for these functions should be opaque for the iterator implementation.