Fix file sharing issue.
If a file was previously opened with FILE_SHARE_WRITE by some other process, the call to CreateFile was failing with ERROR_SHARING_VIOLATION. With this change YARA can read files that were opened with FILE_SHARE_WRITE by other processes.

Thanks to Frank Boldewin for reporting and finding the root cause of this issue.