Fullword fix (#1544)
* Minor re-styling.

* Roll our own is_alnum() equivalent.

Use yr_isalnum() when checking if the match is fullword. See
https://github.com/VirusTotal/yara-python/issues/184 for details of what is
happening here, as this only happens on MacOS python AFAICT.

* Fix word character in RE code.

Use yr_isalnum() in RE code for handling \w.

* Fix styling.

* Roll our own is_alnum() equivalent.

Use yr_isalnum() when checking if the match is fullword. See
https://github.com/VirusTotal/yara-python/issues/184 for details of what is
happening here, as this only happens on MacOS python AFAICT.

* Fix word character in RE code.

Use yr_isalnum() in RE code for handling \w.

* Fix styling.

Co-authored-by: Victor M. Alvarez <vmalvarez@virustotal.com>