Add additional functions to math module (#1483)
* feat: Implement stats module

Add a new statistics module. For now, the module supports counting
the occurrences of specific bytes (or returning the occurrence rate
as a percentage) and calculating the most common byte for an area.

* Another example

* feat: Implement stats module

Add a new statistics module. For now, the module supports counting
the occurrences of specific bytes (or returning the occurrence rate
as a percentage) and calculating the most common byte for an area.

* refactor: Integrate stats module into math module

Integrate the new functionality into the math module instead
of creating a new module.

* chore: Adjust module name in documentation

* chore: Fix memory leak issue

Fixes a memory leak where a temporary array (used for match counting)
wasn't released properly.

* fix: Remove global char count function, optimize char count

Remove a redundant (and slow) function to count a char globally.
Rework the remaining char count function to be more efficient.
It still requires O(n*m) in worst case and O(n) in best case.

* chore: Update documentation

* chore: Remove string counting function

Due to the new "x of them in ..." syntax, string counting is no
longer required (it already is supported and this function is
slower in most circumstances).

* chore: remove test functions for string count

Co-authored-by: Florian Roth <venom14@gmail.com>