Fixes issue #1631.
The lexer doesn't produce an _IDENTIFIER_WITH_WILDCARD_ token anymore, instead it produces two tokens _IDENTIFIER_ and `*`. In the context of a rule_enumeration_item, the grammar handles this sequence of tokens as an identifier with a wildcard instead of a multiplication. The drawback with this solution is that spaces are now valid in between the identifier prefix and the `*`. For example, this is valid:

`all of (some_rule_prefix *) `