Convert is_dotnet() function back to a field.
This is done for consistency with pe.is_pe, which is a field instead of a function. The initial idea was adding the is_pe() function to the pe module, but that's not possible because a function and a field can not have the same name, and is too late to remove the is_pe field.