Fix another compiler edge case. (#1722)
* Fix another compiler edge case.

When compiling and we encounter a range that has values we can determine at
compile time we should error out if the lower bound is less than 0.

* Fix OP_COUNT_IN paranoid check.

The OP_COUNT_IN operation does not need the paranoid check on r1 as it is the
lower bound of the range, and thus a constant and not a pointer somewhere. It
should, instead, be checking r3.

Fixes #1721 (the first half of it).