Increase MAX_PE_EXPORTS to 16384
Increase the MAX_PE_EXPORTS for rules using the PE module, specifically when using pe.exports or pe.number_of_exports, where the exports exceed the current setting of 8192. Its possible to circumvent detecting exports that are defined where the ordinal is greater than 8192, which will prevent YARA rules from detecting the export and the number of exports.