Fixed output of YARA on Windows when output is redirected
According to the documentation of
[_setmode](https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/setmode?view=msvc-170#remarks),
stream should be flushed before it is switched to another mode. This was
causing issues when you redirected output of YARA on Windows to a pipe
or a file. stdout is line buffered when terminal is attached but
once the output is redirected, it's usually buffered by a certain buffer
size. That's how systems usually do it.