Implement bytes keyword.
This commit adds the "bytes" keyword to the language. It can be used to read
arbitrary sequences of bytes similar to the int8 family of keywords. The major
difference here is you can use it to read arbitrary sequences into strings and
compare them.

The implementation is done by dynamically allocating memory for each bytes
operation and tracking all the allocations during bytecode execution in a
notebook and destroying that notebook when execution is done.

An obvious optimization to make here is to track allocation offset and lengths
so that we don't have lots of unnecessary allocations done in loops, like this:

for any s in ("foo", "bar"): (bytes(0, 3) == s)

While the above is an obviously contrived example it does make sense to avoid
unnecessary allocations if we already have them.

Discussed in https://github.com/VirusTotal/yara/issues/1780 by @metthal.