Print module names.
Add a -M option to the cli which will print the module names to stdout. As the
number of modules has grown it can become confusing to know what modules are
available, so this option will display them on stdout.

NOTE: I'm not sure I like how I implemented this. It seems hackish to expose an
API that just prints things on stdout but it felt even more hackish to walk the
yr_modules_table array from outside of the module code. ;)

Fixes #737.