Optimize bytes() allocation logic.
Use a hash table to track allocations by offset and length pairs. When a new
allocation is made we store the pointer to that allocation in the table. If we
receive a request to get the bytes at a duplicate offset and length pair we
return the stored pointer for that allocation and avoid another allocation.

Note: The pointers we are storing are inside a notebook so we do not need to
free them when destroying the table as they are freed when the notebook is
destroyed.