Cleanup allocation design.
Switch away from using a notebook for allocations as they have a limit that
allocations must be no bigger than the page size. While it is possible to work
within this it also has another problem which is that it isn't well suited
to storing SIZED_STRING allocations because they can be various sizes and
notebooks tend to work best when you know the size of the objects being
allocated within the notebook ahead of time.

I'm simplifying the design by just using yr_alloc() directly, but still keeping
the hash table to avoid duplicate allocations. I'm also moving the hash table
into the scan context, which is a better place for it.

Finally, implement limits on the amount of allocations that can be made and the
largest allocation allowed. The total number of allocations is 200 and the
maximum size any one allocation can be is 1KB. These are very conservative
numbers and can be tweaked in the future. In fact, it might be a good idea to
make them runtime configurable.