convert esoteric RDN names from wincrypt to openssl
This commit adds code to convert RDN names from the wincrypt format
to the openssl ones. Only a few are added that could be found in some
certificates used to sign PE files. This can be extended with more
if needed, although a conversion to a hashmap might be useful in that
case.