fix error handling in string.to_int() (#1842)
* fix error handling in string.to_int()

Error handling in the to_int method of the new string module was not
properly done.

- errno was not reset prior to the call to strtoll, so YR_UNDEFINED
  could be returned if a prior call failed, but not the current one.
- overflow/underflow was not properly checked, as it returns LLONG_MIN
  or LLONG_MAX and not 0.

Instead, do what is recommended in the man page for strtol:
- set errno to 0 prior to the call
- check errno after the call

* Flesh out invalid cases for string.to_int

Ensure the behavior of string.to_int is the least surprising and
consistent across all OSes:

- return an error when the string is not entirely parsed (so "10p"
  fails to parse).
- return an error when the string has no digit whatsoever (avoiding
  parsing "foo" as 0).
- remove the evaluation differences on different libc implementations.

Fixes #1843

* validate value of base in string.to_int(string, base)

There seems to be implementation differences on invalid base values,
especially with cygwin. So validate that the base has a valid value
first.