Fix #1851.
This is actually a combination of two bugs. The first bug is that if a string
fits in an atom we never calculated the xor key. The second bug is that if you
specify the xor modifier we were NOT setting the ascii flag on the string, so we
could not calculate the xor key properly one we fixed the first bug.

Tested with:

wxs@mbp yara % cat rules/xor.yara
rule a {
  strings:
    $c = "This" xor
  condition:
    any of them
}
wxs@mbp yara % ./yara -s -X rules/xor.yara tests/data/xorwideandascii.out | head -5
a tests/data/xorwideandascii.out
0x4:$c:xor(0x00,This): This
0x1c:$c:xor(0x01,This): Uihr
0x34:$c:xor(0x02,This): Vjkq
0x4c:$c:xor(0x03,This): Wkjp
wxs@mbp yara %

An interesting side effect of this bug is that you could also trigger it with a
"xor wide" combination of modifiers if the string was short enough to fit in an
atom, which is now also fixed.

wxs@mbp yara % cat rules/xor.yara
rule a {
  strings:
    $c = "Th" xor wide
  condition:
    any of them
}
wxs@mbp yara % ./yara -s -X rules/xor.yara tests/data/xorwideandascii.out | head -5
a tests/data/xorwideandascii.out
0x18f4:$c:xor(0x00,T\x00h\x00): T\x00h\x00
0x191f:$c:xor(0x01,T\x00h\x00): U\x01i\x01
0x194a:$c:xor(0x02,T\x00h\x00): V\x02j\x02
0x1975:$c:xor(0x03,T\x00h\x00): W\x03k\x03
wxs@mbp yara %

Noticed by @melomac, who provided a nice writeup to help point out that this
only happens on "short" strings (which led me to realize it was only on strings
that fit in an atom).