Add support for version_data in rich header.
Pull out just the version data (every other dword) into it's own string. This is
quite useful if you want to hash just the version fields of the rich header,
which makes for a more robust signature because it is now independent of the
count field.

Hashing this value is a useful signature. ;)