Some small fixes in macho module (#1889)
* feat: fix size of yr_arm_thread_state64_t

The last element in the struct is a 32 bit integer. This caused the
struct to be slightly bigger than it should be, possibly causing the
entry point for this architecture to not be retrieved.

* fix macho.entry_point_for_arch if undefined

If the arch does exist in the fat file, but the entry point could not be
retrieved, the entry point is YR_UNDEFINED. Calling
`macho.entry_point_for_arch` on this arch adds this undefined value to
the arch offset, returning a buggy value.