Don't fail when yr_notebook_alloc is called with a large size.
Until now yr_notebook_alloc was assuming that the size of allocated buffer was always less than page size. In the best case, it produced an assertion, but when assertions are turned off with `NDEBUG` this leads to memory corruption. With this change yr_notebook_alloc always fulfil the allocation, creating a larger than normal page if necessary.